pyproject.toml
[tool.poetry]
name = "dbt-ext"
version = "0.0.0"
description = "`dbt-ext` is a Meltano utility extension."
authors = ["Meltano Team and Contributors"]
maintainers = ["Meltano Team and Contributors"]
readme = "README.md"
homepage = "https://meltano.com"
repository = "https://github.com/meltano/dbt-ext"
license = "Apache-2.0"

packages = [
    { include = "dbt_ext" },
]

# If you need to static assets with your extension, you can add them here.
# Under a directory path of "files_dbt-ext/"
include = [
    {path = "files_dbt_ext"},
]

[tool.poetry.dependencies]
python = ">=3.8"
PyYAML = "^6.0.0"
click = "^8.1.3"
typer = "^0.6.1"
"meltano.edk"= { version = "~=0.4.0", python = "<4" }
importlib-resources = "^5.9.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core==1.9.0", "poetry-dynamic-versioning==1.4.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
dbt_extension = 'dbt_ext.main:app'
dbt_invoker = 'dbt_ext.pass_through:pass_through_cli'

[tool.poetry-dynamic-versioning]
enable = true
